Added Hall et al (2016) modifications to model
Hall, Richie and Coppola (2016) is an experimental replication of
Richie et al (2014). I am in the process of modifying the Richie et al
model to reflect (a) the interaction schedule in Hall et al, and (b)
the measure of conventionalization used in Hall et al (jaccard index on
actual utterances).